separated core module into submodules. all tests passing
