+++ tests/insert_behavior_spec.lua
@@ -4,7 +4,8 @@ local test_helpers = require("tests.test_helpers")
 local add_buffer = test_helpers.add_buffer
 
 describe("Streamline insert buffer behavior", function()
-	local core = require("streamline.core")
+	local core = require("streamline.core.core")
+	local buffer_state = require("streamline.core.buffer_state")
 
 	before_each(function()
 		test_helpers.reset_for_test(core)
@@ -14,11 +15,11 @@ describe("Streamline insert buffer behavior", function()
 		core.config.default_insert_behavior = "beginning"
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
 
 		assert.is_true(core.buffers[buf2_id] ~= nil)
 		assert.is_true(core.buffer_order[1] == buf2_id)
@@ -32,18 +33,18 @@ describe("Streamline insert buffer behavior", function()
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
 
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id)
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id)
+		buffer_state:on_buffer_entered(buf3_id)
 
 		assert.is_true(core.buffers[buf3_id] ~= nil)
 		assert.is_true(vim.tbl_contains(core.buffer_order, buf3_id))
@@ -54,11 +55,11 @@ describe("Streamline insert buffer behavior", function()
 		core.config.default_insert_behavior = "end"
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
 
 		assert.is_true(core.buffers[buf2_id] ~= nil)
 		assert.is_true(core.buffer_order[2] == buf2_id)
@@ -72,18 +73,18 @@ describe("Streamline insert buffer behavior", function()
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
 
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id)
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id)
+		buffer_state:on_buffer_entered(buf3_id)
 
 		assert.is_true(core.buffers[buf3_id] ~= nil)
 		assert.is_true(core.buffer_order[3] == buf3_id)
@@ -95,11 +96,11 @@ describe("Streamline insert buffer behavior", function()
 		core.config.default_insert_behavior = "before"
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
 
 		assert.is_true(core.buffers[buf2_id] ~= nil)
 		assert.is_true(core.buffer_order[1] == buf2_id)
@@ -113,18 +114,18 @@ describe("Streamline insert buffer behavior", function()
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
 
-		core:on_buffer_added(buf1_id) -- <test1>
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id) -- <test1>
+		buffer_state:on_buffer_entered(buf1_id)
 
-		core:on_buffer_added(buf2_id) -- <test2><test1.lua>
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id) -- <test2><test1.lua>
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id) -- <test3><test2.lua><test1.lua>
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id) -- <test3><test2.lua><test1.lua>
+		buffer_state:on_buffer_entered(buf3_id)
 
 		assert.is_true(core.buffers[buf3_id] ~= nil)
 		assert.is_true(vim.tbl_contains(core.buffer_order, buf3_id))
@@ -139,18 +140,18 @@ describe("Streamline insert buffer behavior", function()
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
 
-		core:on_buffer_added(buf1_id) -- <test1>
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id) -- <test1>
+		buffer_state:on_buffer_entered(buf1_id)
 
-		core:on_buffer_added(buf2_id) -- <test2><test1.lua>
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id) -- <test2><test1.lua>
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id) -- <test2><test3.lua><test1.lua>
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id) -- <test2><test3.lua><test1.lua>
+		buffer_state:on_buffer_entered(buf3_id)
 
 		assert.is_true(core.buffers[buf3_id] ~= nil)
 		assert.is_true(vim.tbl_contains(core.buffer_order, buf3_id))
@@ -162,11 +163,11 @@ describe("Streamline insert buffer behavior", function()
 		core.config.default_insert_behavior = "after"
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id) -- <test1>
+		buffer_state:on_buffer_added(buf1_id) -- <test1>
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id) -- <test1><test2.lua>
+		buffer_state:on_buffer_added(buf2_id) -- <test1><test2.lua>
 
 		assert.is_true(core.buffers[buf2_id] ~= nil)
 		assert.is_true(core.buffer_order[2] == buf2_id)
@@ -180,18 +181,18 @@ describe("Streamline insert buffer behavior", function()
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
 
-		core:on_buffer_added(buf1_id) -- <test1>
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id) -- <test1>
+		buffer_state:on_buffer_entered(buf1_id)
 
-		core:on_buffer_added(buf2_id) -- <test1><test2.lua>
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id) -- <test1><test2.lua>
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id) -- <test1><test3.lua><test2.lua>
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id) -- <test1><test3.lua><test2.lua>
+		buffer_state:on_buffer_entered(buf3_id)
 
 		assert.is_true(core.buffers[buf3_id] ~= nil)
 		assert.is_true(vim.tbl_contains(core.buffer_order, buf3_id))
@@ -206,18 +207,18 @@ describe("Streamline insert buffer behavior", function()
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
 
-		core:on_buffer_added(buf1_id) -- <test1>
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id) -- <test1>
+		buffer_state:on_buffer_entered(buf1_id)
 
-		core:on_buffer_added(buf2_id) -- <test1><test2.lua>
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id) -- <test1><test2.lua>
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id) -- <test1><test2.lua><test3.lua>
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id) -- <test1><test2.lua><test3.lua>
+		buffer_state:on_buffer_entered(buf3_id)
 
 		assert.is_true(core.buffers[buf3_id] ~= nil)
 		assert.is_true(vim.tbl_contains(core.buffer_order, buf3_id))
