+++ tests/reinsertion_spec.lua
@@ -4,7 +4,9 @@ local test_helpers = require("tests.test_helpers")
 local add_buffer = test_helpers.add_buffer
 
 describe("Streamline reinsert buffers", function()
-	local core = require("streamline.core")
+	local core = require("streamline.core.core")
+	local buffer_state = require("streamline.core.buffer_state")
+	local reordering = require("streamline.core.reordering")
 
 	before_each(function()
 		test_helpers.reset_for_test(core)
@@ -13,15 +15,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 1 before index 1 in a list of 2", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_before_index(1, 1)
+		reordering:reinsert_buffer_before_index(1, 1)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
@@ -31,15 +33,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 1 before index 2 in a list of 2", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_before_index(1, 2)
+		reordering:reinsert_buffer_before_index(1, 2)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
@@ -49,15 +51,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 2 before index 1 in a list of 2", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_before_index(2, 1)
+		reordering:reinsert_buffer_before_index(2, 1)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf2_id)
@@ -67,20 +69,20 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 3 before index 2 in a list of 3", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id)
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id)
+		buffer_state:on_buffer_entered(buf3_id)
 
-		core:reinsert_buffer_before_index(3, 2)
+		reordering:reinsert_buffer_before_index(3, 2)
 
 		assert.is_equal(core.active_buf.id, buf3_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
@@ -91,15 +93,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 2 after index 2 in a list of 2", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_after_index(2, 2)
+		reordering:reinsert_buffer_after_index(2, 2)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
@@ -109,15 +111,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 2 after index 1 in a list of 2", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_after_index(2, 1)
+		reordering:reinsert_buffer_after_index(2, 1)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
@@ -127,15 +129,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 1 after index 2 in a list of 2", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_after_index(1, 2)
+		reordering:reinsert_buffer_after_index(1, 2)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf2_id)
@@ -145,20 +147,20 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 1 after index 2 in a list of 3", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id)
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id)
+		buffer_state:on_buffer_entered(buf3_id)
 
-		core:reinsert_buffer_after_index(1, 2)
+		reordering:reinsert_buffer_after_index(1, 2)
 
 		assert.is_equal(core.active_buf.id, buf3_id)
 		assert.is_equal(core.buffer_order[1], buf2_id)
@@ -169,15 +171,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 1 before index 2 in a list of 2 by id", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_before_id(buf1_id, buf2_id)
+		reordering:reinsert_buffer_before_id(buf1_id, buf2_id)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
@@ -187,15 +189,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 2 before index 1 in a list of 2 by id", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_before_id(buf2_id, buf1_id)
+		reordering:reinsert_buffer_before_id(buf2_id, buf1_id)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf2_id)
@@ -205,20 +207,20 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 3 before index 2 in a list of 3 by id", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id)
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id)
+		buffer_state:on_buffer_entered(buf3_id)
 
-		core:reinsert_buffer_before_id(buf3_id, buf2_id)
+		reordering:reinsert_buffer_before_id(buf3_id, buf2_id)
 
 		assert.is_equal(core.active_buf.id, buf3_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
@@ -229,15 +231,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 2 after index 1 in a list of 2 by id", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_after_id(buf2_id, buf1_id)
+		reordering:reinsert_buffer_after_id(buf2_id, buf1_id)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
@@ -247,15 +249,15 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 1 after index 2 in a list of 2 by id", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:reinsert_buffer_after_id(buf1_id, buf2_id)
+		reordering:reinsert_buffer_after_id(buf1_id, buf2_id)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf2_id)
@@ -265,20 +267,20 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 1 after index 2 in a list of 3 by id", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id)
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id)
+		buffer_state:on_buffer_entered(buf3_id)
 
-		core:reinsert_buffer_after_id(buf1_id, buf2_id)
+		reordering:reinsert_buffer_after_id(buf1_id, buf2_id)
 
 		assert.is_equal(core.active_buf.id, buf3_id)
 		assert.is_equal(core.buffer_order[1], buf2_id)
@@ -289,16 +291,16 @@ describe("Streamline reinsert buffers", function()
 	it("reinsert a buffer at index 1 before index 2 directly after swapping them", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:swap_buffer_with(1, 2)
-		core:reinsert_buffer_before_index(2, 1)
+		reordering:swap_buffer_with(1, 2)
+		reordering:reinsert_buffer_before_index(2, 1)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf1_id)
