+++ tests/swap_spec.lua
@@ -4,7 +4,9 @@ local test_helpers = require("tests.test_helpers")
 local add_buffer = test_helpers.add_buffer
 
 describe("Streamline swap buffer", function()
-	local core = require("streamline.core")
+	local core = require("streamline.core.core")
+	local buffer_state = require("streamline.core.buffer_state")
+	local reordering = require("streamline.core.reordering")
 
 	before_each(function()
 		test_helpers.reset_for_test(core)
@@ -13,14 +15,14 @@ describe("Streamline swap buffer", function()
 	it("swap buffers at index 1 and 2 in a list of 2", function()
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 
-		core:swap_buffer_with(1, 2)
+		reordering:swap_buffer_with(1, 2)
 
 		assert.is_equal(core.active_buf.id, buf2_id)
 		assert.is_equal(core.buffer_order[1], buf2_id)
@@ -32,18 +34,18 @@ describe("Streamline swap buffer", function()
 
 		local buf1_id = add_buffer("test1")
 		test_helpers.track_buffer(buf1_id)
-		core:on_buffer_added(buf1_id)
-		core:on_buffer_entered(buf1_id)
+		buffer_state:on_buffer_added(buf1_id)
+		buffer_state:on_buffer_entered(buf1_id)
 		local buf2_id = add_buffer("test2")
 		test_helpers.track_buffer(buf2_id)
-		core:on_buffer_added(buf2_id)
-		core:on_buffer_entered(buf2_id)
+		buffer_state:on_buffer_added(buf2_id)
+		buffer_state:on_buffer_entered(buf2_id)
 		local buf3_id = add_buffer("test3")
 		test_helpers.track_buffer(buf3_id)
-		core:on_buffer_added(buf3_id)
-		core:on_buffer_entered(buf3_id)
+		buffer_state:on_buffer_added(buf3_id)
+		buffer_state:on_buffer_entered(buf3_id)
 
-		core:swap_buffer_with(1, 3)
+		reordering:swap_buffer_with(1, 3)
 
 		assert.is_equal(core.active_buf.id, buf3_id)
 		assert.is_equal(core.buffer_order[1], buf3_id)
